Organized css style files to new structure
