+++ src/assets/styles/stockdetail.css
