+++ src/assets/styles/stocklist.css
